+++ packages/plugin-packages/mashroom-portal-default-theme/src/backend/dev_server.ts
@@ -9,6 +9,7 @@ import type {Request, Response} from 'express';
 import type {MashroomPortalPageRenderModel} from '@mashroom/mashroom-portal/type-definitions';
 
 themeParams.setParams({
+    showPortalAppHeaders: true,
     showEnvAndVersions: true,
     mashroomVersion: '1.0.0'
 });