+++ packages/plugin-packages/mashroom-portal-default-theme/src/backend/handlebar_helpers.ts
@@ -32,6 +32,14 @@ function mashroomVersion(): string {
     return themeParams.mashroomVersion;
 }
 
+function extraMainClasses(): string {
+    let classes = '';
+    if (themeParams.showPortalAppHeaders) {
+        classes += 'show-portal-app-headers';
+    }
+    return classes;
+}
+
 function ifShowEnvAndVersions(this: any, options: any): any {
     if (themeParams.showEnvAndVersions) {
         return options.fn(this);
@@ -43,6 +51,7 @@ export default {
     equals,
     env,
     mashroomVersion,
+    extraMainClasses,
     ifShowEnvAndVersions,
     '__': i18n,
     defaultPluginErrorMessage,