+++ packages/plugin-packages/mashroom-portal-default-theme/src/backend/theme_params.ts
@@ -1,15 +1,20 @@
 
 type ThemeParams = {
+    showPortalAppHeaders: boolean;
     showEnvAndVersions: boolean;
     mashroomVersion: string;
 }
 
 let params: ThemeParams = {
+    showPortalAppHeaders: true,
     showEnvAndVersions: false,
     mashroomVersion: '',
 };
 
 export default {
+    get showPortalAppHeaders() {
+        return params.showPortalAppHeaders;
+    },
     get showEnvAndVersions() {
         return params.showEnvAndVersions;
     },